type mismatches, prepare for release
git-svn-id: https://svn.r-project.org/R-packages/trunk@6175 694ef91d-65df-0310-b7bb-92e67a308ead

+++ nlme/src/nlmefit.c
@@ -158,7 +158,7 @@ count_DmHalf_pars( dimPTR dd, longint *pdClass )
 	case 3: result += 2;
 	}
     }
-    return (size_t) result;
+    return result;
 }
 
 double *
